+++ lib/src/src/render.dart
@@ -1,12 +1,23 @@
 import 'dart:html' show Element, querySelector;
 
-import 'package:pheasant_temp/pheasant_build.dart' show PheasantTemplate;
+import 'package:pheasant_temp/pheasant_build.dart' show PheasantTemplate, AppState;
 
 /// Function used to encapsulate the injection of the processed [PheasantTemplate] into the DOM
 /// 
-/// With time, the functions here will be developed on further.
-void renderElement(Element elementApp) {
+/// This function injects the [PheasantTemplate] into the DOM, and makes use of the [AppState] to watch for changes in the [app]. 
+/// Whenever a change is emitted, the application is rerendered in the DOM.
+/// 
+/// If [appState] is not provided, then the state is generated in the beginning of the app lifecycle.
+void renderElement(PheasantTemplate app, {AppState? appState}) {
+  AppState state = appState ?? AppState(component: app);
+  PheasantTemplate appObj = app;
+  Element elementApp = appObj.render(appObj.template!, state);
   querySelector('#output')?.children.add(elementApp);
+
+  state.stateStream.listen((event) {
+    elementApp = appObj.render(appObj.template!, state);
+    querySelector('#output')?.children.first = elementApp;
+  },);
 }
 
 /// Function used to create a Pheasant Application
@@ -14,5 +25,5 @@ void renderElement(Element elementApp) {
 /// The function accepts a pheasant file object - [PheasantTemplate] - and then processes the application.
 /// Once it is done, then it is injected into the DOM.
 void createApp(PheasantTemplate pheasantTemplate) {
-  renderElement(pheasantTemplate.render(pheasantTemplate.template!));
+  renderElement(pheasantTemplate);
 }